authentication fix

+++ src/notes/models/noteModel.ts
@@ -2,7 +2,6 @@ import db from "../../db.js";
 import { Note } from "../../types/typeNote.js";
 
 export const getAll = async (userId: number): Promise<Note[]> => {
-  // const result = await db.query("SELECT * FROM notes ORDER BY created_at DESC");
   const result = await db.query("SELECT * FROM notes WHERE user_id = $1", [userId]);
   return result.rows;
 };
@@ -15,22 +14,22 @@ export const create = async (title: string, userId: number): Promise<Note> => {
   return result.rows[0];
 };
 
-export const deleteNote = async (id: number) => {
-  const result = await db.query("DELETE FROM notes WHERE id = $1 RETURNING *", [
-    id,
+export const deleteNote = async (id: number, userId: number) => {
+  const result = await db.query("DELETE FROM notes WHERE id = $1 AND user_id = $2 RETURNING *", [
+    id, userId
   ]);
   return result.rows[0];
 };
 
-export const edit = async (id: number, newTitle: string) => {
+export const edit = async (id: number, newTitle: string, userId: number) => {
   const result = await db.query(
-    "UPDATE notes SET title = $1 WHERE id = $2 RETURNING *",
-    [newTitle, id]
+    "UPDATE notes SET title = $1 WHERE id = $2 AND user_id = $3 RETURNING *",
+    [newTitle, id, userId]
   );
   return result.rows[0];
 };
 
-export const getNoteById = async (id: number) => {
-  const result = await db.query("SELECT * FROM notes WHERE id = $1", [id]);
+export const getNoteById = async (id: number, userId: number) => {
+  const result = await db.query("SELECT * FROM notes WHERE id = $1 AND user_id = $2", [id, userId]);
   return result.rows[0];
 };
